compressor/requirements.txt
boto3==1.16.52
botocore==1.19.52
jmespath==0.10.0
Pillow==8.1.0
PyMySQL==1.0.2
python-dateutil==2.8.1
s3transfer==0.3.4
six==1.15.0
urllib3==1.26.2